requirements.txt
pandas==1.5.3
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2

